# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legal AI Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DocumentStatus =
  (typeof DocumentStatus)[keyof typeof DocumentStatus];

export const DocumentStatus = {
  pending: "pending",
  analyzing: "analyzing",
  analyzed: "analyzed",
  error: "error",
} as const;

export interface Document {
  id: number;
  title: string;
  content: string;
  fileType: string;
  status: DocumentStatus;
  createdAt: string;
  updatedAt: string;
}

export interface CreateDocumentBody {
  title: string;
  content: string;
  fileType: string;
}

export interface Entity {
  /** Entity type: JUDGE, PETITIONER, RESPONDENT, DATE, CASE_NUMBER, COURT, SECTION, ACT */
  type: string;
  value: string;
  confidence: number;
}

export interface LegalProvision {
  section: string;
  act: string;
  title: string;
  relevanceScore: number;
  excerpt: string;
}

export type AnalysisResultOutcomeClassification =
  (typeof AnalysisResultOutcomeClassification)[keyof typeof AnalysisResultOutcomeClassification];

export const AnalysisResultOutcomeClassification = {
  acquittal: "acquittal",
  conviction: "conviction",
  appeal_allowed: "appeal_allowed",
  appeal_dismissed: "appeal_dismissed",
  settlement: "settlement",
  remanded: "remanded",
  unknown: "unknown",
} as const;

export type AnalysisResultBailClassification =
  (typeof AnalysisResultBailClassification)[keyof typeof AnalysisResultBailClassification];

export const AnalysisResultBailClassification = {
  bailable: "bailable",
  non_bailable: "non_bailable",
  unknown: "unknown",
} as const;

export type AnalysisResultCaseType =
  (typeof AnalysisResultCaseType)[keyof typeof AnalysisResultCaseType];

export const AnalysisResultCaseType = {
  civil: "civil",
  criminal: "criminal",
  constitutional: "constitutional",
  family: "family",
  commercial: "commercial",
  unknown: "unknown",
} as const;

export interface AnalysisResult {
  id: number;
  documentId: number;
  summary: string;
  entities: Entity[];
  legalProvisions: LegalProvision[];
  outcomeClassification: AnalysisResultOutcomeClassification;
  bailClassification: AnalysisResultBailClassification;
  caseType: AnalysisResultCaseType;
  verdictConfidence: number;
  keyFindings: string[];
  citedCases: string[];
  createdAt: string;
}

export interface MultiDocumentAnalysisBody {
  documentIds: number[];
}

export type CrossReferenceType =
  (typeof CrossReferenceType)[keyof typeof CrossReferenceType];

export const CrossReferenceType = {
  contradiction: "contradiction",
  pattern: "pattern",
  precedent: "precedent",
  shared_provision: "shared_provision",
} as const;

export type CrossReferenceSeverity =
  (typeof CrossReferenceSeverity)[keyof typeof CrossReferenceSeverity];

export const CrossReferenceSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CrossReference {
  type: CrossReferenceType;
  description: string;
  documentIds: number[];
  severity: CrossReferenceSeverity;
}

export interface MultiDocumentAnalysisResult {
  documentIds: number[];
  crossReferences: CrossReference[];
  commonProvisions: string[];
  overallSummary: string;
  createdAt: string;
}

export interface ChatMessageBody {
  documentId: number;
  message: string;
  /** @nullable */
  sessionId?: string | null;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  id: number;
  documentId: number;
  role: ChatMessageRole;
  content: string;
  sessionId: string;
  createdAt: string;
}

export type DashboardStatsCaseTypeBreakdown = { [key: string]: number };

export type DashboardStatsOutcomeBreakdown = { [key: string]: number };

export interface DashboardStats {
  totalDocuments: number;
  analyzedDocuments: number;
  totalChats: number;
  caseTypeBreakdown: DashboardStatsCaseTypeBreakdown;
  outcomeBreakdown: DashboardStatsOutcomeBreakdown;
  avgVerdictConfidence: number;
}

export interface RecentAnalysis {
  documentId: number;
  documentTitle: string;
  outcomeClassification: string;
  caseType: string;
  createdAt: string;
}