# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Legal AI Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: documents
    description: Legal document management
  - name: analysis
    description: AI legal analysis
  - name: chat
    description: Chatbot interactions
  - name: stats
    description: Dashboard statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Upload and create a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Document created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Invalid input

  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /documents/{id}/analyze:
    post:
      operationId: analyzeDocument
      tags: [analysis]
      summary: Run full AI analysis on a document (RAG + NER + Classifier)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "404":
          description: Not found

  /documents/{id}/analysis:
    get:
      operationId: getDocumentAnalysis
      tags: [analysis]
      summary: Get cached analysis for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "404":
          description: Not found

  /multi-document-analysis:
    post:
      operationId: multiDocumentAnalysis
      tags: [analysis]
      summary: Cross-reference multiple documents for contradictions, patterns, and precedents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MultiDocumentAnalysisBody"
      responses:
        "200":
          description: Multi-document analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultiDocumentAnalysisResult"

  /chat:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message in a chat session about a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageBody"
      responses:
        "200":
          description: Chat response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"

  /chat/{documentId}/history:
    get:
      operationId: getChatHistory
      tags: [chat]
      summary: Get chat history for a document
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"

  /stats/dashboard:
    get:
      operationId: getDashboardStats
      tags: [stats]
      summary: Get dashboard summary statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /stats/recent-analyses:
    get:
      operationId: getRecentAnalyses
      tags: [stats]
      summary: Get recent document analyses
      responses:
        "200":
          description: Recent analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecentAnalysis"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Document:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        fileType:
          type: string
        status:
          type: string
          enum: [pending, analyzing, analyzed, error]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - content
        - fileType
        - status
        - createdAt
        - updatedAt

    CreateDocumentBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        fileType:
          type: string
      required:
        - title
        - content
        - fileType

    Entity:
      type: object
      properties:
        type:
          type: string
          description: "Entity type: JUDGE, PETITIONER, RESPONDENT, DATE, CASE_NUMBER, COURT, SECTION, ACT"
        value:
          type: string
        confidence:
          type: number
      required:
        - type
        - value
        - confidence

    LegalProvision:
      type: object
      properties:
        section:
          type: string
        act:
          type: string
        title:
          type: string
        relevanceScore:
          type: number
        excerpt:
          type: string
      required:
        - section
        - act
        - title
        - relevanceScore
        - excerpt

    AnalysisResult:
      type: object
      properties:
        id:
          type: integer
        documentId:
          type: integer
        summary:
          type: string
        entities:
          type: array
          items:
            $ref: "#/components/schemas/Entity"
        legalProvisions:
          type: array
          items:
            $ref: "#/components/schemas/LegalProvision"
        outcomeClassification:
          type: string
          enum: [acquittal, conviction, appeal_allowed, appeal_dismissed, settlement, remanded, unknown]
        bailClassification:
          type: string
          enum: [bailable, non_bailable, unknown]
        caseType:
          type: string
          enum: [civil, criminal, constitutional, family, commercial, unknown]
        verdictConfidence:
          type: number
        keyFindings:
          type: array
          items:
            type: string
        citedCases:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - documentId
        - summary
        - entities
        - legalProvisions
        - outcomeClassification
        - bailClassification
        - caseType
        - verdictConfidence
        - keyFindings
        - citedCases
        - createdAt

    MultiDocumentAnalysisBody:
      type: object
      properties:
        documentIds:
          type: array
          items:
            type: integer
      required:
        - documentIds

    CrossReference:
      type: object
      properties:
        type:
          type: string
          enum: [contradiction, pattern, precedent, shared_provision]
        description:
          type: string
        documentIds:
          type: array
          items:
            type: integer
        severity:
          type: string
          enum: [low, medium, high]
      required:
        - type
        - description
        - documentIds
        - severity

    MultiDocumentAnalysisResult:
      type: object
      properties:
        documentIds:
          type: array
          items:
            type: integer
        crossReferences:
          type: array
          items:
            $ref: "#/components/schemas/CrossReference"
        commonProvisions:
          type: array
          items:
            type: string
        overallSummary:
          type: string
        createdAt:
          type: string
      required:
        - documentIds
        - crossReferences
        - commonProvisions
        - overallSummary
        - createdAt

    ChatMessageBody:
      type: object
      properties:
        documentId:
          type: integer
        message:
          type: string
        sessionId:
          type: ["string", "null"]
      required:
        - documentId
        - message

    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        documentId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        sessionId:
          type: string
        createdAt:
          type: string
      required:
        - id
        - documentId
        - role
        - content
        - sessionId
        - createdAt

    DashboardStats:
      type: object
      properties:
        totalDocuments:
          type: integer
        analyzedDocuments:
          type: integer
        totalChats:
          type: integer
        caseTypeBreakdown:
          type: object
          additionalProperties:
            type: integer
        outcomeBreakdown:
          type: object
          additionalProperties:
            type: integer
        avgVerdictConfidence:
          type: number
      required:
        - totalDocuments
        - analyzedDocuments
        - totalChats
        - caseTypeBreakdown
        - outcomeBreakdown
        - avgVerdictConfidence

    RecentAnalysis:
      type: object
      properties:
        documentId:
          type: integer
        documentTitle:
          type: string
        outcomeClassification:
          type: string
        caseType:
          type: string
        createdAt:
          type: string
      required:
        - documentId
        - documentTitle
        - outcomeClassification
        - caseType
        - createdAt