# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Legal AI Platform API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all documents
 */
export const ListDocumentsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  fileType: zod.string(),
  status: zod.enum(["pending", "analyzing", "analyzed", "error"]),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListDocumentsResponse = zod.array(ListDocumentsResponseItem);

/**
 * @summary Upload and create a document
 */
export const CreateDocumentBody = zod.object({
  title: zod.string(),
  content: zod.string(),
  fileType: zod.string(),
});

/**
 * @summary Get a document by ID
 */
export const GetDocumentParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDocumentResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  fileType: zod.string(),
  status: zod.enum(["pending", "analyzing", "analyzed", "error"]),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a document
 */
export const DeleteDocumentParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Run full AI analysis on a document (RAG + NER + Classifier)
 */
export const AnalyzeDocumentParams = zod.object({
  id: zod.coerce.number(),
});

export const AnalyzeDocumentResponse = zod.object({
  id: zod.number(),
  documentId: zod.number(),
  summary: zod.string(),
  entities: zod.array(
    zod.object({
      type: zod
        .string()
        .describe(
          "Entity type: JUDGE, PETITIONER, RESPONDENT, DATE, CASE_NUMBER, COURT, SECTION, ACT",
        ),
      value: zod.string(),
      confidence: zod.number(),
    }),
  ),
  legalProvisions: zod.array(
    zod.object({
      section: zod.string(),
      act: zod.string(),
      title: zod.string(),
      relevanceScore: zod.number(),
      excerpt: zod.string(),
    }),
  ),
  outcomeClassification: zod.enum([
    "acquittal",
    "conviction",
    "appeal_allowed",
    "appeal_dismissed",
    "settlement",
    "remanded",
    "unknown",
  ]),
  bailClassification: zod.enum(["bailable", "non_bailable", "unknown"]),
  caseType: zod.enum([
    "civil",
    "criminal",
    "constitutional",
    "family",
    "commercial",
    "unknown",
  ]),
  verdictConfidence: zod.number(),
  keyFindings: zod.array(zod.string()),
  citedCases: zod.array(zod.string()),
  createdAt: zod.string(),
});

/**
 * @summary Get cached analysis for a document
 */
export const GetDocumentAnalysisParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDocumentAnalysisResponse = zod.object({
  id: zod.number(),
  documentId: zod.number(),
  summary: zod.string(),
  entities: zod.array(
    zod.object({
      type: zod
        .string()
        .describe(
          "Entity type: JUDGE, PETITIONER, RESPONDENT, DATE, CASE_NUMBER, COURT, SECTION, ACT",
        ),
      value: zod.string(),
      confidence: zod.number(),
    }),
  ),
  legalProvisions: zod.array(
    zod.object({
      section: zod.string(),
      act: zod.string(),
      title: zod.string(),
      relevanceScore: zod.number(),
      excerpt: zod.string(),
    }),
  ),
  outcomeClassification: zod.enum([
    "acquittal",
    "conviction",
    "appeal_allowed",
    "appeal_dismissed",
    "settlement",
    "remanded",
    "unknown",
  ]),
  bailClassification: zod.enum(["bailable", "non_bailable", "unknown"]),
  caseType: zod.enum([
    "civil",
    "criminal",
    "constitutional",
    "family",
    "commercial",
    "unknown",
  ]),
  verdictConfidence: zod.number(),
  keyFindings: zod.array(zod.string()),
  citedCases: zod.array(zod.string()),
  createdAt: zod.string(),
});

/**
 * @summary Cross-reference multiple documents for contradictions, patterns, and precedents
 */
export const MultiDocumentAnalysisBody = zod.object({
  documentIds: zod.array(zod.number()),
});

export const MultiDocumentAnalysisResponse = zod.object({
  documentIds: zod.array(zod.number()),
  crossReferences: zod.array(
    zod.object({
      type: zod.enum([
        "contradiction",
        "pattern",
        "precedent",
        "shared_provision",
      ]),
      description: zod.string(),
      documentIds: zod.array(zod.number()),
      severity: zod.enum(["low", "medium", "high"]),
    }),
  ),
  commonProvisions: zod.array(zod.string()),
  overallSummary: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Send a message in a chat session about a document
 */
export const SendChatMessageBody = zod.object({
  documentId: zod.number(),
  message: zod.string(),
  sessionId: zod.string().nullish(),
});

export const SendChatMessageResponse = zod.object({
  id: zod.number(),
  documentId: zod.number(),
  role: zod.enum(["user", "assistant"]),
  content: zod.string(),
  sessionId: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Get chat history for a document
 */
export const GetChatHistoryParams = zod.object({
  documentId: zod.coerce.number(),
});

export const GetChatHistoryResponseItem = zod.object({
  id: zod.number(),
  documentId: zod.number(),
  role: zod.enum(["user", "assistant"]),
  content: zod.string(),
  sessionId: zod.string(),
  createdAt: zod.string(),
});
export const GetChatHistoryResponse = zod.array(GetChatHistoryResponseItem);

/**
 * @summary Get dashboard summary statistics
 */
export const GetDashboardStatsResponse = zod.object({
  totalDocuments: zod.number(),
  analyzedDocuments: zod.number(),
  totalChats: zod.number(),
  caseTypeBreakdown: zod.record(zod.string(), zod.number()),
  outcomeBreakdown: zod.record(zod.string(), zod.number()),
  avgVerdictConfidence: zod.number(),
});

/**
 * @summary Get recent document analyses
 */
export const GetRecentAnalysesResponseItem = zod.object({
  documentId: zod.number(),
  documentTitle: zod.string(),
  outcomeClassification: zod.string(),
  caseType: zod.string(),
  createdAt: zod.string(),
});
export const GetRecentAnalysesResponse = zod.array(
  GetRecentAnalysesResponseItem,
);